Remove anon functions from cli
