+++ cli/key.go
@@ -6,14 +6,12 @@ package cli
 import (
 	"context"
 
-	"github.com/ava-labs/avalanchego/ids"
-
-	"github.com/ava-labs/hypersdk/codec"
+	"github.com/ava-labs/hypersdk/cli/prompt"
 	"github.com/ava-labs/hypersdk/rpc"
 	"github.com/ava-labs/hypersdk/utils"
 )
 
-func (h *Handler) SetKey(lookupBalance func(int, string, string, uint32, ids.ID) error) error {
+func (h *Handler) SetKey() error {
 	keys, err := h.GetKeys()
 	if err != nil {
 		return err
@@ -37,13 +35,22 @@ func (h *Handler) SetKey(lookupBalance func(int, string, string, uint32, ids.ID)
 	}
 	utils.Outf("{{cyan}}stored keys:{{/}} %d\n", len(keys))
 	for i := 0; i < len(keys); i++ {
-		if err := lookupBalance(i, h.c.Address(keys[i].Address), uris[0], networkID, chainID); err != nil {
+		addrStr := h.c.Address(keys[i].Address)
+		balance, err := h.c.LookupBalance(addrStr, uris[0], networkID, chainID)
+		if err != nil {
 			return err
 		}
+		utils.Outf(
+			"%d) {{cyan}}address:{{/}} %s {{cyan}}balance:{{/}} %s %s\n",
+			i,
+			addrStr,
+			utils.FormatBalance(balance, h.c.Decimals()),
+			h.c.Symbol(),
+		)
 	}
 
 	// Select key
-	keyIndex, err := h.PromptChoice("set default key", len(keys))
+	keyIndex, err := prompt.Choice("set default key", len(keys))
 	if err != nil {
 		return err
 	}
@@ -51,7 +58,7 @@ func (h *Handler) SetKey(lookupBalance func(int, string, string, uint32, ids.ID)
 	return h.StoreDefaultKey(key.Address)
 }
 
-func (h *Handler) Balance(checkAllChains bool, promptAsset bool, printBalance func(codec.Address, string, uint32, ids.ID, ids.ID) error) error {
+func (h *Handler) Balance(checkAllChains bool) error {
 	addr, _, err := h.GetDefaultKey(true)
 	if err != nil {
 		return err
@@ -59,13 +66,6 @@ func (h *Handler) Balance(checkAllChains bool, promptAsset bool, printBalance fu
 	chainID, uris, err := h.GetDefaultChain(true)
 	if err != nil {
 		return err
-	}
-	var assetID ids.ID
-	if promptAsset {
-		assetID, err = h.PromptAsset("assetID", true)
-		if err != nil {
-			return err
-		}
 	}
 
 	max := len(uris)
@@ -79,9 +79,17 @@ func (h *Handler) Balance(checkAllChains bool, promptAsset bool, printBalance fu
 		if err != nil {
 			return err
 		}
-		if err := printBalance(addr, uri, networkID, chainID, assetID); err != nil {
+		addrStr := h.c.Address(addr)
+		balance, err := h.c.LookupBalance(addrStr, uris[0], networkID, chainID)
+		if err != nil {
 			return err
 		}
+		utils.Outf(
+			"{{cyan}}address:{{/}} %s {{cyan}}balance:{{/}} %s %s\n",
+			addrStr,
+			utils.FormatBalance(balance, h.c.Decimals()),
+			h.c.Symbol(),
+		)
 	}
 	return nil
 }
