+++ cli/spam.go
@@ -18,10 +18,12 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/ava-labs/hypersdk/chain"
+	"github.com/ava-labs/hypersdk/cli/prompt"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/consts"
 	"github.com/ava-labs/hypersdk/fees"
@@ -50,11 +52,48 @@ var (
 	sent     atomic.Int64
 )
 
+type SpamHelper interface {
+	// CreateAccount generates a new account and returns the [PrivateKey].
+	//
+	// The spammer tracks all created accounts and orchestrates the return of funds
+	// sent to any created accounts on shutdown. If the spammer exits ungracefully,
+	// any funds sent to created accounts will be lost unless they are persisted by
+	// the [SpamHelper] implementation.
+	CreateAccount() (*PrivateKey, error)
+	// GetFactory returns the [chain.AuthFactory] for a given private key.
+	//
+	// A [chain.AuthFactory] signs transactions and provides a unit estimate
+	// for using a given private key (needed to estimate fees for a transaction).
+	GetFactory(pk *PrivateKey) (chain.AuthFactory, error)
+
+	// CreateClient instructs the [SpamHelper] to create and persist a VM-specific
+	// JSONRPC client.
+	//
+	// This client is used to retrieve the [chain.Parser] and the balance
+	// of arbitrary addresses.
+	//
+	// TODO: consider making these functions part of the required JSONRPC
+	// interface for the HyperSDK.
+	CreateClient(uri string, networkID uint32, chainID ids.ID) error
+	GetParser(ctx context.Context) (chain.Parser, error)
+	LookupBalance(choice int, address string) (uint64, error)
+
+	// GetTransfer returns a list of actions that sends [amount] to a given [address].
+	//
+	// Memo is used to ensure that each transaction is unique (even if between the same
+	// sender and receiver for the same amount).
+	GetTransfer(address codec.Address, amount uint64, memo []byte) []chain.Action
+}
+
 func (h *Handler) Spam(sh SpamHelper) error {
 	ctx := context.Background()
 
 	// Select chain
-	chainID, uris, err := h.PromptChain("select chainID", nil)
+	chains, err := h.GetChains()
+	if err != nil {
+		return err
+	}
+	chainID, uris, err := prompt.SelectChain("select chainID", chains)
 	if err != nil {
 		return err
 	}
@@ -81,7 +120,7 @@ func (h *Handler) Spam(sh SpamHelper) error {
 		}
 		balances[i] = balance
 	}
-	keyIndex, err := h.PromptChoice("select root key", len(keys))
+	keyIndex, err := prompt.Choice("select root key", len(keys))
 	if err != nil {
 		return err
 	}
@@ -109,34 +148,34 @@ func (h *Handler) Spam(sh SpamHelper) error {
 	}
 
 	// Collect parameters
-	numAccounts, err := h.PromptInt("number of accounts", consts.MaxInt)
+	numAccounts, err := prompt.Int("number of accounts", consts.MaxInt)
 	if err != nil {
 		return err
 	}
 	if numAccounts < 2 {
 		return ErrInsufficientAccounts
 	}
-	sZipf, err := h.PromptFloat("s (Zipf distribution = [(v+k)^(-s)], Default = 1.01)", consts.MaxFloat64)
+	sZipf, err := prompt.Float("s (Zipf distribution = [(v+k)^(-s)], Default = 1.01)", consts.MaxFloat64)
 	if err != nil {
 		return err
 	}
-	vZipf, err := h.PromptFloat("v (Zipf distribution = [(v+k)^(-s)], Default = 2.7)", consts.MaxFloat64)
+	vZipf, err := prompt.Float("v (Zipf distribution = [(v+k)^(-s)], Default = 2.7)", consts.MaxFloat64)
 	if err != nil {
 		return err
 	}
-	txsPerSecond, err := h.PromptInt("txs to try and issue per second", consts.MaxInt)
+	txsPerSecond, err := prompt.Int("txs to try and issue per second", consts.MaxInt)
 	if err != nil {
 		return err
 	}
-	minTxsPerSecond, err := h.PromptInt("minimum txs to issue per second", consts.MaxInt)
+	minTxsPerSecond, err := prompt.Int("minimum txs to issue per second", consts.MaxInt)
 	if err != nil {
 		return err
 	}
-	txsPerSecondStep, err := h.PromptInt("txs to increase per second", consts.MaxInt)
+	txsPerSecondStep, err := prompt.Int("txs to increase per second", consts.MaxInt)
 	if err != nil {
 		return err
 	}
-	numClients, err := h.PromptInt("number of clients per node", consts.MaxInt)
+	numClients, err := prompt.Int("number of clients per node", consts.MaxInt)
 	if err != nil {
 		return err
 	}
@@ -262,7 +301,7 @@ func (h *Handler) Spam(sh SpamHelper) error {
 						float64(confirmedTxs)/float64(totalTxs)*100,
 						inflight.Load(),
 						current-psent,
-						ParseDimensions(unitPrices),
+						unitPrices,
 					)
 				}
 				l.Unlock()
