+++ examples/morpheusvm/cmd/morpheus-cli/cmd/chain.go
@@ -4,14 +4,7 @@
 package cmd
 
 import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/ids"
 	"github.com/spf13/cobra"
-
-	"github.com/ava-labs/hypersdk/chain"
-
-	brpc "github.com/ava-labs/hypersdk/examples/morpheusvm/rpc"
 )
 
 var chainCmd = &cobra.Command{
@@ -65,9 +58,6 @@ var chainInfoCmd = &cobra.Command{
 var watchChainCmd = &cobra.Command{
 	Use: "watch",
 	RunE: func(_ *cobra.Command, args []string) error {
-		return handler.Root().WatchChain(hideTxs, func(uri string, networkID uint32, chainID ids.ID) (chain.Parser, error) {
-			cli := brpc.NewJSONRPCClient(uri, networkID, chainID)
-			return cli.Parser(context.TODO())
-		}, handleTx)
+		return handler.Root().WatchChain(hideTxs)
 	},
 }
