+++ examples/morpheusvm/cmd/morpheus-cli/cmd/key.go
@@ -4,10 +4,8 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
 
-	"github.com/ava-labs/avalanchego/ids"
 	"github.com/spf13/cobra"
 
 	"github.com/ava-labs/hypersdk/auth"
@@ -18,8 +16,6 @@ import (
 	"github.com/ava-labs/hypersdk/crypto/secp256r1"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
 	"github.com/ava-labs/hypersdk/utils"
-
-	brpc "github.com/ava-labs/hypersdk/examples/morpheusvm/rpc"
 )
 
 const (
@@ -186,47 +182,16 @@ var importKeyCmd = &cobra.Command{
 	},
 }
 
-func lookupSetKeyBalance(choice int, address string, uri string, networkID uint32, chainID ids.ID) error {
-	// TODO: just load once
-	cli := brpc.NewJSONRPCClient(uri, networkID, chainID)
-	balance, err := cli.Balance(context.TODO(), address)
-	if err != nil {
-		return err
-	}
-	addr, err := codec.ParseAddressBech32(consts.HRP, address)
-	if err != nil {
-		return err
-	}
-	keyType, err := getKeyType(addr)
-	if err != nil {
-		return err
-	}
-	utils.Outf(
-		"%d) {{cyan}}address (%s):{{/}} %s {{cyan}}balance:{{/}} %s %s\n",
-		choice,
-		keyType,
-		address,
-		utils.FormatBalance(balance, consts.Decimals),
-		consts.Symbol,
-	)
-	return nil
-}
-
 var setKeyCmd = &cobra.Command{
 	Use: "set",
 	RunE: func(*cobra.Command, []string) error {
-		return handler.Root().SetKey(lookupSetKeyBalance)
+		return handler.Root().SetKey()
 	},
-}
-
-func lookupKeyBalance(addr codec.Address, uri string, networkID uint32, chainID ids.ID, _ ids.ID) error {
-	_, err := handler.GetBalance(context.TODO(), brpc.NewJSONRPCClient(uri, networkID, chainID), addr)
-	return err
 }
 
 var balanceKeyCmd = &cobra.Command{
 	Use: "balance",
 	RunE: func(*cobra.Command, []string) error {
-		return handler.Root().Balance(checkAllChains, false, lookupKeyBalance)
+		return handler.Root().Balance(checkAllChains)
 	},
 }
